fix double

+++ database/migrations/2018_03_19_104534_create_transactions_table.php
@@ -14,7 +14,7 @@ class CreateTransactionsTable extends Migration
     public function up()
     {
         Schema::table('transactions', function (Blueprint $table) {
-            $table->double('sale_amount')->nullable()->change();
+            $table->float('sale_amount')->nullable()->change();
         });
     }
